BIG refactor logic
