+++ tests/test_url_converter.py
@@ -1,16 +1,15 @@
-# pylint: disable=unused-argument
-
-from wb.cloud_agent.main import get_controller_url
-from wb.cloud_agent.settings import AppSettings, base_url_to_agent_url
+from wb.cloud_agent.settings import AppSettings
+from wb.cloud_agent.utils import get_controller_url
 
 
-def test_base_url_to_agent_url():
+def test_base_url_to_agent_url(settings: AppSettings):
     assert (
-        base_url_to_agent_url("https://cloud-staging.wirenboard.com/")
+        settings.base_url_to_agent_url("https://cloud-staging.wirenboard.com/")
         == "https://agent.cloud-staging.wirenboard.com/api-agent/v1/"
     )
     assert (
-        base_url_to_agent_url("https://wirenboard.cloud/") == "https://agent.wirenboard.cloud/api-agent/v1/"
+        settings.base_url_to_agent_url("https://wirenboard.cloud/")
+        == "https://agent.wirenboard.cloud/api-agent/v1/"
     )
