+++ wb/cloud_agent/utils.py
@@ -1,0 +1,95 @@
+import json
+import logging
+import subprocess
+import sys
+from functools import cache
+from pathlib import Path
+from typing import TYPE_CHECKING
+from urllib.parse import urljoin
+
+from tabulate import tabulate
+
+if TYPE_CHECKING:
+    from wb.cloud_agent.settings import Provider
+
+
+@cache
+def get_ctrl_serial_number() -> str:
+    return subprocess.check_output("wb-gen-serial -s", shell=True).decode().strip()
+
+
+def get_controller_url(base_url: str) -> str:
+    ctrl_serial_number = get_ctrl_serial_number()
+    return urljoin(base_url, f"controllers/{ctrl_serial_number}")
+
+
+def read_json_config(config_path: Path) -> dict[str, str]:
+    data = config_path.read_text(encoding="utf-8")
+    try:
+        return json.loads(data)
+    except json.JSONDecodeError:
+        print(f"Error parsing JSON in: {config_path}")
+        sys.exit(6)
+
+
+def read_plaintext_config(config_path: Path) -> str:
+    with config_path.open("r", encoding="utf-8") as f:
+        return f.readline().strip()
+
+
+def write_to_file(fpath: Path, contents: str) -> None:
+    fpath.parent.mkdir(parents=True, exist_ok=True)
+    fpath.write_text(contents, encoding="utf-8")
+
+
+def start_and_enable_service(service: str, restart: bool = False, timeout: int = 120) -> None:
+    logging.debug("Enabling service %s", service)
+
+    result = subprocess.run(
+        ["systemctl", "enable", service],
+        check=True,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        text=True,
+        timeout=timeout,
+    )
+    if result.stdout:
+        logging.debug("Enabling service stdout: %s", result.stdout.strip())
+    if result.stderr:
+        logging.debug("Enabling service stderr: %s", result.stderr.strip())
+
+    if restart:
+        logging.debug("Restarting service %s", service)
+        subprocess.run(["systemctl", "restart", service], check=True, timeout=timeout)
+    else:
+        logging.debug("Starting service %s", service)
+        subprocess.run(["systemctl", "start", service], check=True, timeout=timeout)
+
+
+def stop_and_disable_service(service: str, timeout: int = 120) -> None:
+    logging.debug("Stopping service %s", service)
+    subprocess.run(["systemctl", "stop", service], check=True, timeout=timeout)
+
+    logging.debug("Disabling service %s", service)
+    result = subprocess.run(
+        ["systemctl", "disable", service],
+        check=True,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        text=True,
+        timeout=timeout,
+    )
+    if result.stdout:
+        logging.debug("Disabling service stdout: %s", result.stdout.strip())
+    if result.stderr:
+        logging.debug("Disabling service stderr: %s", result.stderr.strip())
+
+
+def show_providers_table(providers: list["Provider"]) -> None:
+    if not providers:
+        print("No one provider was found")
+        return
+
+    table = [[p.name, p.display_url] for p in providers]
+    headers = ["Provider", "Controller Url / Activation Url"]
+    print(tabulate(table, headers=headers, tablefmt="github"))
